# openapi.yaml
openapi: 3.1.0
info:
  title: Qaida Places API
  version: 1.0.0
  summary: API inferred from the Qaida mobile UI kit screens.
  description: |
    This specification was derived from the HTML screens in the design folder.
    It describes the backend contract implied by the search, recommendations,
    place details, favorites, profile, collections, and notifications flows.

    Some parts are inferred from UI behavior and should be confirmed before implementation:
    - recommendation scoring logic
    - booking flows mentioned in notifications
    - collection vs default favorites behavior
    - push notification delivery strategy
servers:
  - url: https://api.qaida.app/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /places:
    get:
      tags: [Places]
      summary: List places
      description: Returns places filtered for category, budget, company type, distance, and sort order.
      parameters:
        - $ref: '#/components/parameters/Category'
        - $ref: '#/components/parameters/Budget'
        - $ref: '#/components/parameters/CompanyType'
        - $ref: '#/components/parameters/DistanceKm'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Place list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /places/popular:
    get:
      tags: [Places]
      summary: Get popular places
      description: Returns a curated list used on home and onboarding screens.
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Popular places
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceListResponse'
  /places/nearby:
    get:
      tags: [Places]
      summary: Get nearby places
      parameters:
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/DistanceKm'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Nearby places
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /places/{placeId}:
    get:
      tags: [Places]
      summary: Get place details
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '200':
          description: Detailed place information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /places/{placeId}/menu:
    get:
      tags: [Places]
      summary: Get place menu
      parameters:
        - $ref: '#/components/parameters/PlaceId'
        - in: query
          name: category
          schema:
            type: string
          description: Optional menu category filter.
      responses:
        '200':
          description: Menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
                required: [items]
        '404':
          $ref: '#/components/responses/NotFound'
  /places/{placeId}/reviews:
    get:
      tags: [Reviews]
      summary: List reviews for a place
      parameters:
        - $ref: '#/components/parameters/PlaceId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Reviews]
      summary: Create review for a place
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /places/{placeId}/share:
    post:
      tags: [Places]
      summary: Generate share link for a place
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '200':
          description: Share link generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /places/{placeId}/complaints:
    post:
      tags: [Places]
      summary: Submit complaint about a place
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '202':
          description: Complaint accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /search:
    get:
      tags: [Search]
      summary: Search places
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Category'
        - $ref: '#/components/parameters/Budget'
        - $ref: '#/components/parameters/CompanyType'
        - $ref: '#/components/parameters/DistanceKm'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Search results with optional match score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /recommendations:
    get:
      tags: [Recommendations]
      summary: Get recommendations using query filters
      parameters:
        - $ref: '#/components/parameters/CompanyType'
        - $ref: '#/components/parameters/Category'
        - $ref: '#/components/parameters/Budget'
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Recommended places
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationListResponse'
    post:
      tags: [Recommendations]
      summary: Create personalized recommendations from explicit preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecommendationRequest'
      responses:
        '200':
          description: Personalized recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PlaceId:
      in: path
      name: placeId
      required: true
      schema:
        type: string
        format: uuid
...
    Category:
      in: query
      name: category
      schema:
        type: array
        items:
          type: string
      style: form
      explode: false
    Budget:
      in: query
      name: budget
      schema:
        $ref: '#/components/schemas/BudgetRange'
    CompanyType:
      in: query
      name: companyType
      schema:
        $ref: '#/components/schemas/CompanyType'
    DistanceKm:
      in: query
      name: distanceKm
      schema:
        type: number
        format: float
        minimum: 0
    SortBy:
      in: query
      name: sortBy
      schema:
        type: string
        enum: [popularity, rating, distance, recent]
    Latitude:
      in: query
      name: lat
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    Longitude:
      in: query
      name: lng
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    BudgetRange:
      type: string
      enum: ['$', '$$', '$$$']
    CompanyType:
      type: string
      enum: [solo, couple, friends, family]
    PlaceStatus:
      type: string
      enum: [open, closed, coming_soon]
...
    Coordinates:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
      required: [lat, lng]
...
    PlaceSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        reviewCount:
          type: integer
          minimum: 0
        address:
          type: string
        distanceKm:
          type: number
          format: float
        budget:
          $ref: '#/components/schemas/BudgetRange'
        status:
          $ref: '#/components/schemas/PlaceStatus'
        imageUrl:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string
        isFavorite:
          type: boolean
      required:
        [id, name, category, rating, reviewCount, address, budget, status, isFavorite]
    PlaceDetail:
      allOf:
        - $ref: '#/components/schemas/PlaceSummary'
        - type: object
          properties:
            description:
              type: string
            coordinates:
              $ref: '#/components/schemas/Coordinates'
            openingHours:
              type: array
              items:
                $ref: '#/components/schemas/OpeningHour'
            gallery:
              type: array
              items:
                type: string
                format: uri
            amenities:
              type: array
              items:
                type: string
            menuPreview:
              type: array
              items:
                $ref: '#/components/schemas/MenuItem'
          required: [description]
    OpeningHour:
      type: object
      properties:
        day:
          type: string
        open:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
        close:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
      required: [day, open, close]
    MenuItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        currency:
          type: string
          default: KZT
        imageUrl:
          type: string
          format: uri
      required: [id, name, category, price]
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        placeId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        authorName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, placeId, userId, authorName, rating, text, createdAt]
    Recommendation:
      type: object
      properties:
        place:
          $ref: '#/components/schemas/PlaceSummary'
        matchScore:
          type: integer
          minimum: 0
          maximum: 100
        reasons:
          type: array
          items:
            type: string
      required: [place, matchScore]
...
    CreateRecommendationRequest:
      type: object
      properties:
        companyType:
          $ref: '#/components/schemas/CompanyType'
        categories:
          type: array
          items:
            type: string
        budget:
          $ref: '#/components/schemas/BudgetRange'
        coordinates:
          $ref: '#/components/schemas/Coordinates'
      required: [companyType]
    CreateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
          minLength: 1
      required: [rating, text]
...
    CreateComplaintRequest:
      type: object
      properties:
        reason:
          type: string
        description:
          type: string
      required: [reason]
    ShareResponse:
      type: object
      properties:
        shareUrl:
          type: string
          format: uri
      required: [shareUrl]
    AcceptedResponse:
      type: object
      properties:
        accepted:
          type: boolean
          const: true
      required: [accepted]
    PlaceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlaceSummary'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
    ReviewListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
    RecommendationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
      required: [items]
    SearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        total:
          type: integer
          minimum: 0
      required: [items, total]
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
...